Fixed random forest error for large sample sizes
The random forest method errored out when given a large sample sizes.
Setting proximity=FALSE in both instances avoides the creation of an n x
n matrix.
This should fix the error.
Output from the function with TRUE and FALSE for proximity was compared
and is identical in the exampleData tutorial results.